Fixed query string not being passed along to unmocked endpoints with superagent 3.6+
superagent clears `this._query` in `_finalizeQueryString` by setting it's length
to 0, for idempotency.  Then, after superagent-mock decides it is not interested
in the request, it tries to call `originalEnd()`, which would run the real
request without the query.

Fixed by taking a clone of the original `this._query` value before calling
`_finalizeQueryString`, then resetting the property afterwards as is done
with `this.url`.